[Integration: Test] refactor spring cloud
