+++ shenyu-integrated-test/shenyu-integrated-test-spring-cloud/docker-compose.yml
@@ -21,15 +21,11 @@ services:
     restart: always
     networks:
       - shenyu
-    volumes:
-      - type: bind
-        source: ./src/main/resources/admin-config/
-        target: /opt/sql/
     ports:
       - 9095:9095
     environment:
       - SPRING_PROFILES_ACTIVE=h2
-      - "shenyu.database.init_script=sql-script/h2/schema.sql;file:/opt/sql/schema.sql"
+      - "shenyu.database.init_script=sql-script/h2/schema.sql"
     healthcheck:
       test: ["CMD-SHELL", "wget -q -O - http://shenyu-admin:9095/actuator/health | grep UP || exit 1"]
       timeout: 2s
@@ -52,9 +48,6 @@ services:
       - 8761:8761
     networks:
       - shenyu
-    depends_on:
-      shenyu-admin:
-        condition: service_healthy
   shenyu-examples-springcloud:
     deploy:
       resources:
@@ -76,8 +69,8 @@ services:
     networks:
       - shenyu
     depends_on:
-      shenyu-integrated-test-spring-cloud:
-          condition: service_healthy
+      - shenyu-examples-eureka
+      - shenyu-admin
   #  from http plugin dockerfile
   shenyu-integrated-test-spring-cloud:
     container_name: shenyu-integrated-test-spring-cloud
